rework auth into api/gatsby functions, install netlify plugin for vault, remove lambda redirect, update megaprofile test
