--- lambda/config/indexMedia.config.js
@@ -1,20 +1,0 @@
-module.exports = {
-  itunes: {
-    podcasts: [
-      // Classes Without Quizzes at Reunion Homecoming
-      385577095,
-      // Stanford Alumni Stories Project
-      557942848,
-    ],
-  },
-  soundcloud: {
-    podcastFeeds: [
-      "https://feeds.simplecast.com/r4JJTrrW",
-      "https://feeds.soundcloud.com/users/soundcloud:users:357652052/sounds.rss",
-      "https://feeds.soundcloud.com/users/soundcloud:users:146499962/sounds.rss",
-    ],
-  },
-  youtube: {
-    channelId: ["UCNoz7zRzEcHhp3TBU6spaXA"],
-  },
-};
