+++ netlify.toml
@@ -2,7 +2,6 @@
 # ##############################################################################
 [build]
   publish = "public"
-  functions = "lambda"
 
 [dev]
   publish = "public"
@@ -15,7 +14,7 @@
   for = "/*"
   [headers.values]
     Access-Control-Allow-Origin = "*"
-    Content-Security-Policy = "default-src https: 'unsafe-inline' 'unsafe-eval'; form-action https:; img-src https: data:"
+#    Content-Security-Policy = "default-src https: 'unsafe-inline' 'unsafe-eval'; form-action https:; img-src https: data:"
     X-Content-Type-Options = "nosniff"
     Referrer-Policy = "strict-origin-when-cross-origin"
     Strict-Transport-Security = "max-age=2592000"
@@ -43,7 +42,7 @@
 # Fetch vault plugin needs to come before any of the contextual environment
 # variable plugins so they can act on the new vars.
 [[plugins]]
-  package = "/plugins/netlify-plugin-vault-variables"
+  package = "netlify-plugin-vault-variables"
   [plugins.inputs]
     endpoint = 'https://vault.stanford.edu'
     # Order doesn't matter here as these will be fetch asyncronously. Try not to have dupes.
@@ -51,12 +50,7 @@
     # Set which environment variables to use for the approle login
     secretId = 'VAULT_SECRET_ID'
     roleId = 'VAULT_ROLE_ID'
-
-[[plugins]]
-  package = 'netlify-plugin-contextual-env'
-  [plugins.inputs]
-    mode = 'prefix'
-
+  
 [[plugins]]
   package = '/plugins/netlify-plugin-contextual-auth'
   [plugins.inputs]
@@ -65,18 +59,13 @@
     context = 'branch-deploy'
     branch = 'staging'
 
-# [[plugins]]
-# package = "/plugins/netlify-plugin-wildcard-404-redirect"
-#  [plugins.inputs]
-#  destination = "https://cardinalalumni.stanford.edu/:splat"
+[[plugins]]
+package = "/plugins/netlify-plugin-wildcard-404-redirect"
+ [plugins.inputs]
+ destination = "https://cardinalalumni.stanford.edu/:splat"
 
 [[plugins]]
 package = "@netlify/plugin-gatsby"
-
-[[plugins]]
-package = "netlify-plugin-inline-functions-env"
-  [plugins.inputs]
-  buildEvent = "onBuild"
 
 # CONTEXTUAL PLUGINS
 # ##############################################################################
