+++ package.json
@@ -42,6 +42,8 @@
     "gatsby-source-storyblok": "^3.0.1",
     "luxon": "^1.27.0",
     "netlify-plugin-inline-functions-env": "^1.0.9",
+    "netlify-plugin-vault-variables": "^1.0.4",
+    "next-connect": "^0.12.2",
     "path-browserify": "^1.0.1",
     "postscribe": "^2.0.8",
     "process": "^0.11.10",
@@ -83,7 +85,8 @@
     "start": "npm run develop",
     "serve": "gatsby serve",
     "clean": "gatsby clean",
-    "vault": "node ./plugins/netlify-plugin-vault-variables/script.js",
+    "vault": "netlify-plugin-vault-variables",
+    "vault:local": "NODE_ENV=development netlify-plugin-vault-variables",
     "lint": "NODE_ENV=test eslint \"./src/**/*.{js,jsx,ts,tsx}\"",
     "lint:fix": "NODE_ENV=test eslint \"./src/**/*.{js,jsx,ts,tsx}\" --fix",
     "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\" && exit 1",
