--- plugins/netlify-plugin-vault-variables/manifest.yml
@@ -1,17 +1,0 @@
-# manifest.yml
-
-name: netlify-plugin-vault-variables
-inputs:
-  - name: endpoint
-    description: Which vault endpoint to query
-    required: true
-  - name: secrets
-    description: An array of strings that are the vault path for the secrets.
-    required: true
-    default: 'string'
-  - name: secretId
-    description: Vault secret ID for approle login
-    required: true
-  - name: roleId
-    description: Vault role ID for approle login
-    required: true
