+++ src/api/auth/logout.js
@@ -1,0 +1,3 @@
+import { authInstance } from '../../utilities/authInstance';
+
+export default authInstance.destroySession();
