+++ static/_redirects
@@ -5,9 +5,6 @@
 # This redirect is here because the vanity url service does not allow for multipart subdomains.
 https://trusteenominations.alumni.stanford.edu https://alumni.stanford.edu/volunteer/trustee-nominations 301!
 
-# WILDCARD FOR API LAMBDA FUNCTIONS
-################################################################################
-/api/auth/* /.netlify/functions/auth/:splat 200
 
 # PATH REDIRECTS
 ################################################################################
